config.h -> Uncomment TEMP_RESIDENCY_TIME and TEMP_HYSTERESIS defines and set to 10 seconds and 2 degC respectively. This to see if it helps with print quality issues seen on the initial layers of prints. Upper layers seem to stabilise and produce better quality / more consistent output during initial development testing. Suspect it could be a temperature issue with the ABS plastic. The use of TEMP_RESIDENCY_TIME will ensure that when the wait for extruder temperature command is issued the printer will wait for the temperature to stabilise to the target for at least TEMP_RESIDENCY_TIME (set to 10 seconds for now) before continuing to print.
TEMP_HYSTERESIS is also used to provide an acceptable target temperature band before M109 returns and allows the print to proceed. Previously, the temperature can sometimes hover 1 or 2 degC above target and not be able to begin printing unless the extruder head was 'blown' on to reduce the temperature below the target.