Version 1.3.23i-VCP.
Added M852 command to allow entry in to the DFU bootloader via the firmware application. This will allow the Makibox firmware to be updated without the need to open up the Makibox and press the reset button on the PrintRBoard or mess around with removing / reconnecting jumpers. A 'F' pass code must also be entered together with the M852 command. This is to prevent accidental entry in to the bootloader by end-users.