Version 13.23l-VCP
Uncomment WATCHPERIOD. This enables checks of the hotend heater when a target temperature is set. If the temperature does not appear to be increasing after WATCHPERIOD has elapsed then the target temperature is set to 0 degC in order to protect the hotend heater. This situation may occur if the hotend thermistor has not been installed at the right location and therefore not measuring the hotend temperature.

Small edit to the WATCHPERIOD checking logic in heater.c