Version 1.3.23s-VCP
Edit format of M109 and M190 (wait for heater to reach temperature) text so that Pronterface will process it and display the temperature on the GUI graph if 'Monitor Printer' is selected. Added a timeout to M109 and M190 while loops.

config.h - Increased DROP_SEGMENTS to 50. _AXIS_STEP_PER_UNIT changed to 240 for the new extruder drive - needs to be double checked with final drive design.

Added BED_HEATUP_TIMEOUT and HOTEND_HEATUP_TIMEOUT in config.h