Version 1.3.24i-VCP
Ability to track the positions of the X, Y, Z, and E axes added. This is based on counting the actual numer of steps issued to the stepper driver. The system must be homed first for accurate position feedback. On first power up, the existing positions are assumed to be the zero position (X, Y, and Z will not display negative positions - zero is the minimum).

M114 now uses a new function in makibox.c to output text feedack for the position.

Slight update to the make file.