Version 1.3.24y-VCP
Added code in stepper.c to turn off x/y/z stepper motors if they are not required for the current 'block'. This is in order to reduce the power dissipation (heating) in the stepper motors.

This code has been copied from the 'pause-print' development branch to the main / default branch since this is a feature request for the near future.