Version 1.3.25i-VCP
Merged the 'default' branch and the 'pause-resume' development branch.

Note SRAM usage is currently rather high. Bear this in mind when adding further features and adding more variables / strings / arrays. You may run in to stack overflow problems. Reducing the plan buffer size may be a way to go if difficulties are enountered BUT the plan buffer size was set to 64 to prevent it from being completely run dry when there are a series of relatively short move.