Version 1.3.25m-VCP
Added support for M18 (to turn off / disable motors). This operates in exactly the same way as M84 which is already implemented. Users have reported that the OctoPrint host software sends a M18 command when trying the stop / disable motors. Pronterface / PrintRun sends a M84 when the 'Motors Off' button is pressed.

NOTE: Please remember that the bed thermistor table is for the newer hot-bed where the thermistor has been located on the bottom of the board and positioned at the centre.