Version 1.3.25q-VCP
SD Card Development Branch

Initial development branch to get the SD card working. This is an initial commit to the repository with first small steps and experimenting with a SD card library (open source GNU v2 license) from Roland Riegel (www.roland-riegel.de). Many thanks to Roland Riegel for making this available as open source under the GNU GPL v2 and LGPL V2.1.

So far in this version of code we have been able to successfully initialise a SD card (i.e. the SPI module is working) and read basic information such as capacity and free space available from the SD card.

Expecting further issues with the Y-STOP / SS input pin on the PrintRBoard. This is used as an input for the Y-axis end-stop. However, this is also the SS (slave select) pin for the SPI module. When this input is low it puts the SPI module in to 'slave' mode. The Atmel documentation does not show any means of disabling this 'feature'. It will most likely cause us problems unless we find a reliable work around.