Version 1.3.25r-VCP
SD Card Dev Branch

Makefile - Changed compiler optimisation level to -O1 (simple optimisation). This appears to resolve the crash / memory issues that was seen when testing this code => or has this simply moved the problem to a differene memory location? Had a search through the code for any pointer arithmetic that may go wrong but could not spot anything.

Can now list files and folders from the root directory of the SD card. Related code has just been added to the bottom of the makibox.c file for testing purposes. If the functions are to be used in a release then they should seperated out to a different file (and added to the project) in order for easier navigation and maintenance of the code in the future.

This is still a very early work in progress for the SD card functionality.