Version 1.3.25w-VCP
SD Card Dev Branch

Correct bootloader jump address in JumpToBootloader(void) (within makibox.c.).
Corrected to 0x1E000 (byte address). The 0xF000 address listed in the datasheet
is the word address.

Now able to upload / write to the SD card via Pronterface. The transfer / upload
seems a bit slower than desired. This may be due to, at least in part, the fact
that Pronterface waits until an OK response is received from the printer after
each data line it sends before sending the next line.

M30 Delete SD card file feature implemented and tested using a terminal. It was
not possible to test this command in Pronterface as there are no menus with this
option and the Pronterface command line terminal converts all letters to capitals
meaning that it would not work.