Version 2.03a / 22.09.2013
Another attempt to add the makibox_sdcard.h file to repository as it seems that
it was not added properly in the last commit / repository push.

Steps per unit for the extruder drive updated to 165 as a result of testing in
the lab showing that a setting of 154 was slightly under-extruding during prints.
Hopefully, 165 will provide a better starting point for the majority of users.