Version 2.04 / 05.10.2013
Steps per unit for the extruder drive updated to 163 as a result of testing in
the lab.

Added language.h to collect all the text sent over the USB connection to a list
of #defines. This enables easier translation of this text if desired in the
future. Note that some items of text are read / used by some host software such
as Pronterface and Repetier. Therefore, these English versions of the text and
the format must be maintained were necessary in order to maintain compatibility
with these host software.